+++ src/rule/one_of_rule.cpp
@@ -16,11 +16,11 @@ bool OneOfRule::validate(const Variant &target, ValidationContext &context) cons
 	}
 
 	// Exactly one sub-rule must pass
-	int passed_count = 0;
+	int64_t passed_count = 0;
 	std::vector<size_t> passed_indices;
 	std::vector<ValidationContext> failed_contexts;
 
-	for (size_t i = 0; i < sub_rules.size(); i++) {
+	for (int64_t i = 0; i < sub_rules.size(); i++) {
 		ValidationContext sub_context = context.create_child_schema(vformat("oneOf/%d", i));
 
 		if (sub_rules[i]->validate(target, sub_context)) {
@@ -34,9 +34,7 @@ bool OneOfRule::validate(const Variant &target, ValidationContext &context) cons
 	}
 
 	if (passed_count == 0) {
-		context.add_error(
-				vformat("Value failed all %d oneOf schemas", (int)sub_rules.size()),
-				"oneOf");
+		context.add_error(vformat("Value failed all %d oneOf schemas", (int64_t)sub_rules.size()), "oneOf");
 
 		// Show errors from first failed schema for debugging
 		if (!failed_contexts.empty()) {
@@ -54,9 +52,8 @@ bool OneOfRule::validate(const Variant &target, ValidationContext &context) cons
 			passed_list += String::num(passed_indices[i]);
 		}
 
-		context.add_error(
-				vformat("Value matched %d oneOf schemas (indices: %s) but exactly 1 is required",
-						passed_count, passed_list),
+		context.add_error(vformat("Value matched %d oneOf schemas (indices: %s) but exactly 1 is required",
+								  passed_count, passed_list),
 				"oneOf");
 		return false;
 	}
